refactor: Move `Song` class to `tracks.py`
